feat(ui): Add ExperimentalMaterial3Api annotation to CurrentStatusCard
The code changes in this commit add the `@ExperimentalMaterial3Api` annotation to the `CurrentStatusCard` composable function in the file `CurrentStatusCard.kt`. This annotation marks the usage of experimental Material 3 API.

chore(ui): Refactor ToggleButton in LocationScreen

This commit refactors the `RoutingButton` composable function in the file `LocationScreen.kt`. The function has been renamed to `ToggleButton`, and it now accepts an additional optional parameter for customizing the button text. The purpose of this change is to improve code readability and maintainability.